Changed while loop to for loop, added renderNewTweets function and interval
The for loop will be easier to reuse later when wrapped in a function. The renderNewTweets function stores new tweets in a variable and will call the for loop. This will be called every 1.5 seconds if the update variable is true. Users will be able to toggle the update variable